Create a Form folder and move Form component into it
