+++ front/src/components/Form/Form.jsx
@@ -7,9 +7,7 @@ function Form() {
         password: '',
     });
 
-    const [errors, setErrors] = useState({
-        
-    });
+    const [errors, setErrors] = useState({});
 
     function handleChange(evento) {
         setUserData({
